refactor(article-fields): add any fields from rest_framework to another file
